requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
bibtexparser==1.4.3
blinker==1.9.0
blis==1.2.0
build==1.2.2.post1
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.21.0
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.2.0
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docutils==0.21.2
durationpy==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
faiss-cpu==1.10.0
fake-useragent==2.0.3
fastapi==0.115.8
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.1.24
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.66.0
googlesearch-python==1.3.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langcodes==3.5.0
langsmith==0.3.4
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.25.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.61.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
posthog==3.11.0
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
pypdf==5.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2025.1
PyYAML==6.0.2
ragas==0.2.12
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
roman-numerals-py==3.1.0
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scholarly==1.7.11
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.29.0
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.2.1
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.37
srsly==2.5.1
starlette==0.45.3
streamlit==1.41.1
sympy==1.13.3
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
trio==0.29.0
trio-websocket==0.12.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
#uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
